Extend importTableList file syntax with negations, wildcards, custom SELECTS.
msdbsc support only currently. Documentation and installer/packages have not
been updated yet. Other flavors are broken for the time being.

Instead of a simple list of table names, files given as an importTableList
are treated as a list of rules. Each table is checked against each rule.
A rule is either a regular rule or a negation. Negations begin with a "-".
The last rule to match a table in the target database wins. If the rule
is not a negation, the table is imported. Rules may be simple table names
(optionally including schema) or may contain wildcards "*".

A rule of * is treated as *.*. If there are only negative rules, there is
an implicit * rule at the top.

Rules may specify a custom SELECT statement. This only makes sense
for rules that specify an un-negated literal table name with no wildcards
but is syntactically valid in rules that are negated or have wildcards.

Custom select syntax looks like:

myschema.mytable : SELECT * FROM mytable WHERE import = 1

Calculating which tables to import is now O(nm) where n is the
number of tables and m is the number of rules. In practice this is not a
performance problem unless there are thousands of tables and rules.