Don't copy npgsql Mono.Security.dll to bin on mono builds.
Mono builds should use the system Mono.Security.dll instead of the one
that comes with npgsql. But mono insists on copying the Mono.Security
dll from the npgsql package folder even if the pgdbsc.csproj does not
reference Mono.Security at all, simply because npgsql references it
and it is there. So remove the dll after the build when not building on
windows.